pyproject.toml
[project]
name = "media-pipe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "certifi>=2025.1.31",
    "chromadb>=1.0.9",
    "clip>=0.2.0",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.10.0",
    "google-generativeai>=0.8.4",
    "loadenv>=0.1.1",
    "mediapipe>=0.10.21",
    "neo4j>=5.28.1",
    "openai>=1.82.0",
    "protobuf>=4.25.6",
    "py2neo>=2021.2.4",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=4.1.0",
    "torchfile>=0.1.0",
    "ultralytics>=8.3.93",
    "wandb>=0.19.9",
]
